requirements.txt
Django==5.1
PyPDF2==3.0.1
python-magic==0.4.27
reportlab==4.0.7
pillow==11.0.0
whitenoise==6.8.2
djangorestframework==3.14.0
